config/requirements.txt
msgpack-python==0.4.8
PyYAML==3.11
